+++ src/main/java/com/develetter/develetter/blog/scheduler/BlogSchedulerImpl.java
@@ -5,8 +5,8 @@ import com.develetter.develetter.blog.entity.Blog;
 import com.develetter.develetter.blog.repository.BlogRepository;
 import com.develetter.develetter.blog.service.InterestService;
 import com.develetter.develetter.blog.service.SearchService;
-import com.develetter.develetter.jobposting.UserFilter;
-import com.develetter.develetter.jobposting.UserFilterRepository;
+import com.develetter.develetter.userfilter.entity.UserFilter;
+import com.develetter.develetter.userfilter.repository.UserFilterRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
