+++ src/main/java/com/develetter/develetter/userfilter/repository/UserFilterRepository.java
@@ -1,5 +1,6 @@
-package com.develetter.develetter.jobposting;
+package com.develetter.develetter.userfilter.repository;
 
+import com.develetter.develetter.userfilter.entity.UserFilter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
